LF-4788 Relocate mock out of test utils
