+++ packages/api/tests/utils/ensembleUtils.ts
@@ -15,12 +15,7 @@
 
 import mocks from '../mock.factories.js';
 import { ENSEMBLE_BRAND } from '../../src/util/ensemble.js';
-import { EsciReturnedPrescriptionDetails } from '../../src/util/ensembleService.types.js';
 import { AddonPartner, Farm } from '../../src/models/types.js';
-import LocationModel from '../../src/models/locationModel.js';
-import ManagementPlanModel from '../../src/models/managementPlanModel.js';
-import { getStartOfDate } from '../../src/util/date.js';
-import { customError } from '../../src/util/customErrors.js';
 
 export const connectFarmToEnsemble = async (farm: Farm, partner?: AddonPartner) => {
   const [farmAddon] = await mocks.farm_addonFactory({
@@ -31,99 +26,4 @@ export const connectFarmToEnsemble = async (farm: Farm, partner?: AddonPartner) 
   });
 
   return { farmAddon };
-};
-
-/* Reverse the logic generating the mock id to pull a datetime from it */
-export const getDateTimeFromDayOfMonth = (day: number): Date => {
-  const now = new Date();
-  const date = new Date(now.getFullYear(), now.getMonth(), day);
-
-  return getStartOfDate(date);
-};
-
-export const generateMockPrescriptionDetails = async (
-  farm_id: string,
-  ip_id: number,
-): Promise<EsciReturnedPrescriptionDetails | undefined> => {
-  const locations = await LocationModel.getCropSupportingLocationsByFarmId(farm_id);
-
-  if (locations.length === 0) {
-    throw customError('No crop supporting locations found for the farm', 404);
-  }
-
-  const managementPlans = await ManagementPlanModel.getManagementPlansByLocationId(
-    locations[0].location_id,
-  );
-
-  // Use frontend mock's random toggle between URI and VRI
-  // generate the zones if VRI -- use Duncan's function
-
-  return {
-    id: ip_id,
-    location_id: locations[0].location_id,
-    management_plan_id: managementPlans[0]?.management_plan_id ?? null,
-    recommended_start_datetime: getDateTimeFromDayOfMonth(ip_id).toISOString(),
-    pivot: {
-      center: {
-        // TODO -- use Duncan's function
-        lat: 49.06547,
-        lng: -123.15597,
-      },
-      radius: 150,
-    },
-    metadata: {
-      weather_forecast: {
-        temperature: 20,
-        temperature_unit: '˚C',
-        wind_speed: 10,
-        wind_speed_unit: 'm/s',
-        cumulative_rainfall: 5,
-        cumulative_rainfall_unit: 'mm',
-        et_rate: 2,
-        et_rate_unit: 'mm/h',
-        weather_icon_code: '02d',
-      },
-    },
-    estimated_time: 2,
-    estimated_time_unit: 'h',
-    prescription: {
-      vriData: {
-        file_url: '<https://example.com/vri_data.json>',
-        zones: [
-          {
-            soil_moisture_deficit: 50,
-            application_depth: 15,
-            application_depth_unit: 'mm',
-            grid_points: [
-              {
-                lat: 49.06682,
-                lng: -123.15597,
-              },
-              // ...
-              {
-                lat: 49.0663,
-                lng: -123.1563,
-              },
-            ],
-          },
-          {
-            soil_moisture_deficit: 60,
-            application_depth: 20,
-            application_depth_unit: 'mm',
-            grid_points: [
-              {
-                lat: 49.06479,
-                lng: -123.15776,
-              },
-              // ...
-              {
-                lat: 49.0654,
-                lng: -123.1575,
-              },
-            ],
-          },
-        ],
-      },
-    },
-  };
 };
